add interfaces

+++ apps/user-api/src/adapters/http-adapter/src/controllers/users.controller.ts
@@ -1,5 +1,15 @@
-import { Controller, Get, Post, Body, Put, Delete } from '@nestjs/common';
-import { UserDto } from '../dto/user.dto';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Put,
+  Delete,
+  BadRequestException,
+} from '@nestjs/common';
+import { UpdateUserDto } from '../dto/update-user.dto';
+import { CreateUserDto } from '../dto/create-user.dto';
+import { DeleteUserDto } from '../dto/delete-user.dto';
 import { ApiResponse, ApiTags } from '@nestjs/swagger';
 import {
   CreateUserResult,
@@ -15,7 +25,7 @@ export class UsersController {
   constructor(private readonly userService: UsersService) {}
   @Post()
   @ApiResponse({ status: 201 })
-  async createUser(@Body() dto: UserDto): Promise<CreateUserResult> {
+  async createUser(@Body() dto: CreateUserDto): Promise<CreateUserResult> {
     return await this.userService.createUser(dto);
   }
   @Get()
@@ -23,11 +33,14 @@ export class UsersController {
     return await this.userService.findAll();
   }
   @Delete()
-  async deleteUserById(@Body() dto: UserDto): Promise<DeleteUserResult> {
+  async deleteUserById(@Body() dto: DeleteUserDto): Promise<DeleteUserResult> {
+    if (!dto || dto.id === undefined) {
+      throw new BadRequestException('Не передан id пользователя');
+    }
     return await this.userService.deleteUserById(dto);
   }
   @Put()
-  async updateUserById(@Body() dto: UserDto): Promise<UpdateUserResult> {
+  async updateUserById(@Body() dto: UpdateUserDto): Promise<UpdateUserResult> {
     return await this.userService.updateUserById(dto);
   }
 }
